pyproject.toml
[tool.poetry]
name = "drf-letters"
version = "0.1.0"
description = ""
authors = ["Sergei Sorochinskii <sergei.sorochinskii@dteam.dev>"]
readme = "README.md"
packages = [{include = "drf_letters"}]

[tool.poetry.dependencies]
python = "^3.10"
django-cors-headers = "^3.13.0"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
Markdown = "^3.4.1"
django-filter = "^22.1"
psycopg2-binary = "^2.9.5"
PyJWT = "^2.6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
